+++ packages/assets-controllers/src/NftController.ts
@@ -2065,10 +2065,8 @@ export class NftController extends BaseController<
    * Reset the controller state to the initial state.
    */
   resetState() {
-    this.update((state) => {
-      state.allNftContracts = {};
-      state.allNfts = {};
-      state.ignoredNfts = [];
+    this.update(() => {
+      return getDefaultNftControllerState();
     });
   }
 }